web_app/requirements.txt
feature_engine==1.5.2
matplotlib==3.5.1
numpy==1.21.0
pandas==1.4.2
path.py==12.5.0
scikit_learn==1.1.0
seaborn==0.11.2
streamlit==1.12.2
